requirements.txt
blinker==1.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.7
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.9.1
numpy==1.24.4
pdfminer.six==20250324
pycparser==2.22
pymysql==1.1.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
sqlalchemy==2.0.43
threadpoolctl==3.5.0
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
